[feature/entity] user, report 엔터티 추가 (#3)
- user, report 엔터티 추가
- user 테이블 CRUD 테스트 API 생성

+++ src/users/users.service.ts
@@ -21,15 +21,15 @@ export class UsersService {
     return this.usersRepository.find();
   }
 
-  async findOne(id: number) {
-    const user = await this.usersRepository.findOneBy({ id });
+  async findOne(id: string) {
+    const user = await this.usersRepository.findOneBy({ userId: id });
 
     if (!user) throw new NotFoundException('User not found');
 
     return user;
   }
 
-  async remove(id: number) {
-    await this.usersRepository.delete(id);
+  async remove(id: string) {
+    await this.usersRepository.delete({ userId: id });
   }
 }
